requirements.txt
# Django Framework
Django==6.0.2

# PostgreSQL Database Adapter
psycopg2-binary==2.9.10

# Environment Variable Management
python-decouple==3.8

# Database URL Parser
dj-database-url==2.2.0

# Django REST Framework
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.1
